extract testing axis types into microcrate
