+++ crates/bitnet-bdd-grid-core/Cargo.toml
@@ -13,6 +13,7 @@ description = "Core BDD scenario, feature, and grid primitives for BitNet"
 include = ["src/**", "Cargo.toml"]
 
 [dependencies]
+bitnet-testing-axis-core.workspace = true
 
 [dev-dependencies]
 proptest.workspace = true
