+++ crates/bitnet-testing-axis-core/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "bitnet-testing-axis-core"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Core testing scenario and execution environment axes for BitNet"
+include = ["src/**", "Cargo.toml"]
+
+[dependencies]
